ui: red error messages, texting, status at bottom, F2 now quits
Error messages should be colored in red and will be stored as seperate
buffer. You can send and receive messages. The chat history is also
shown on screen. Lines that span multiple lines are not drawn correctly.
The status messages are now at the bottom. F2 now quits the program (was
`q').