net: packet encryption using serpent
After a salt message has been exchanged, all ongoing communication is
encrypted to prevent eavesdropping. This breaks authentication failure
because it probably yields a protocol error and it does not link this to
an authentication failure.

The mouse cursor is also repositioned to maintain a consistent and
correct cursor focus.