ui: text history no overlapping anymore, increased log size
Every time a text message is added to the history the tables are
recomputed to make sure it fits in the screen. The keyboard wait delay
has been decreased in order to increase responsivity.

wrapaddstr now takes care of strings that may be too long to fit in the
current window. The screen will be cleared when the terminal is resized
in order to get rid of graphical artifacts.